backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
deepgram-sdk==3.0.0
exceptiongroup==1.3.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.6.2
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
packaging==25.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.8.0
requests==2.31.0
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.14.0
urllib3==2.4.0
verboselogs==1.7
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
